refactor: extract technician base status scoring helpers
